car_service_chatbot/requirements.txt
flask==2.0.1
flask-cors==3.0.10
openai==0.27.8
python-dotenv==0.19.0
pandas==1.3.5
numpy==1.21.6
